+++ src/test/java/mil/nga/geopackage/test/extension/elevation/ElevationTilesPngImportTest.java
@@ -7,8 +7,8 @@ import mil.nga.geopackage.BoundingBox;
 import mil.nga.geopackage.core.srs.SpatialReferenceSystem;
 import mil.nga.geopackage.core.srs.SpatialReferenceSystemDao;
 import mil.nga.geopackage.extension.elevation.ElevationTileResults;
-import mil.nga.geopackage.extension.elevation.ElevationTilesPng;
 import mil.nga.geopackage.extension.elevation.ElevationTilesAlgorithm;
+import mil.nga.geopackage.extension.elevation.ElevationTilesPng;
 import mil.nga.geopackage.projection.Projection;
 import mil.nga.geopackage.projection.ProjectionConstants;
 import mil.nga.geopackage.projection.ProjectionFactory;
@@ -30,6 +30,7 @@ public class ElevationTilesPngImportTest extends
 		ImportElevationTilesGeoPackageTestCase {
 
 	private static final boolean PRINT = false;
+	private static final boolean allowNulls = false;
 
 	/**
 	 * Test the elevation extension with a newly created GeoPackage using the
@@ -39,7 +40,7 @@ public class ElevationTilesPngImportTest extends
 	public void testElevationsNearestNeighbor() throws Exception {
 
 		ElevationTilesPngTestUtils.testElevations(geoPackage, null,
-				ElevationTilesAlgorithm.NEAREST_NEIGHBOR, false);
+				ElevationTilesAlgorithm.NEAREST_NEIGHBOR, allowNulls);
 
 	}
 
@@ -51,7 +52,7 @@ public class ElevationTilesPngImportTest extends
 	public void testElevationsBilinear() throws Exception {
 
 		ElevationTilesPngTestUtils.testElevations(geoPackage, null,
-				ElevationTilesAlgorithm.BILINEAR, false);
+				ElevationTilesAlgorithm.BILINEAR, allowNulls);
 
 	}
 
@@ -63,7 +64,7 @@ public class ElevationTilesPngImportTest extends
 	public void testElevationsBicubic() throws Exception {
 
 		ElevationTilesPngTestUtils.testElevations(geoPackage, null,
-				ElevationTilesAlgorithm.BICUBIC, false);
+				ElevationTilesAlgorithm.BICUBIC, allowNulls);
 
 	}
 
@@ -197,7 +198,7 @@ public class ElevationTilesPngImportTest extends
 	@Test
 	public void testBounds() throws Exception {
 
-		long geoPackageEpsg = ProjectionConstants.EPSG_WEB_MERCATOR;
+		long requestEpsg = ProjectionConstants.EPSG_WEB_MERCATOR;
 
 		double widthPixelDistance = 1000;
 		double heightPixelDistance = 1000;
@@ -211,7 +212,7 @@ public class ElevationTilesPngImportTest extends
 		BoundingBox boundingBox = new BoundingBox(minLongitude, maxLongitude,
 				minLatitude, maxLatitude);
 
-		Projection projection = ProjectionFactory.getProjection(geoPackageEpsg);
+		Projection projection = ProjectionFactory.getProjection(requestEpsg);
 		Projection printProjection = ProjectionFactory
 				.getProjection(ProjectionConstants.EPSG_WORLD_GEODETIC_SYSTEM);
 		ProjectionTransform wgs84Transform = projection
@@ -273,36 +274,44 @@ public class ElevationTilesPngImportTest extends
 				}
 				for (double lon = minLongitude + (widthPixelDistance * .5); lon <= maxLongitude; lon += widthPixelDistance) {
 					Double elevation = ElevationTilesPngTestUtils.getElevation(
-							geoPackage, algorithm, lat, lon, geoPackageEpsg);
+							geoPackage, algorithm, lat, lon, requestEpsg);
 					if (PRINT) {
 						System.out.print("   " + elevation);
 					}
-					TestCase.assertNotNull(elevation);
+					if (!allowNulls) {
+						TestCase.assertNotNull(elevation);
+					}
 				}
 			}
 
 			ElevationTileResults results = ElevationTilesPngTestUtils
 					.getElevations(geoPackage, algorithm, boundingBox, width,
-							height, geoPackageEpsg);
-			TestCase.assertNotNull(results);
-			if (PRINT) {
-				System.out.println();
-				System.out.println();
-				System.out.println(algorithm.name());
+							height, requestEpsg);
+			if (!allowNulls) {
+				TestCase.assertNotNull(results);
 			}
-			Double[][] elevations = results.getElevations();
-			TestCase.assertEquals(height, elevations.length);
-			TestCase.assertEquals(width, elevations[0].length);
-			for (int y = 0; y < elevations.length; y++) {
+			if (results != null) {
 				if (PRINT) {
 					System.out.println();
+					System.out.println();
+					System.out.println(algorithm.name());
 				}
-				for (int x = 0; x < elevations[0].length; x++) {
-					Double elevation = elevations[y][x];
+				Double[][] elevations = results.getElevations();
+				TestCase.assertEquals(height, elevations.length);
+				TestCase.assertEquals(width, elevations[0].length);
+				for (int y = 0; y < elevations.length; y++) {
 					if (PRINT) {
-						System.out.print("   " + elevation);
+						System.out.println();
 					}
-					TestCase.assertNotNull(elevation);
+					for (int x = 0; x < elevations[0].length; x++) {
+						Double elevation = elevations[y][x];
+						if (PRINT) {
+							System.out.print("   " + elevation);
+						}
+						if (!allowNulls) {
+							TestCase.assertNotNull(elevation);
+						}
+					}
 				}
 			}
 		}
@@ -318,16 +327,8 @@ public class ElevationTilesPngImportTest extends
 	@Test
 	public void testFullBoundingBox() throws Exception {
 
-		long geoPackageEpsg = ProjectionConstants.EPSG_WEB_MERCATOR;
-
 		int width = 10;
 		int height = 6;
-
-		Projection projection = ProjectionFactory.getProjection(geoPackageEpsg);
-		Projection printProjection = ProjectionFactory
-				.getProjection(ProjectionConstants.EPSG_WORLD_GEODETIC_SYSTEM);
-		ProjectionTransform wgs84Transform = projection
-				.getTransformation(printProjection);
 
 		List<String> elevationTables = ElevationTilesPng.getTables(geoPackage);
 		TileMatrixSetDao dao = geoPackage.getTileMatrixSetDao();
@@ -335,6 +336,16 @@ public class ElevationTilesPngImportTest extends
 		for (String elevationTable : elevationTables) {
 
 			TileMatrixSet tileMatrixSet = dao.queryForId(elevationTable);
+
+			long geoPackageEpsg = tileMatrixSet.getSrs()
+					.getOrganizationCoordsysId();
+
+			Projection projection = ProjectionFactory
+					.getProjection(geoPackageEpsg);
+			Projection printProjection = ProjectionFactory
+					.getProjection(ProjectionConstants.EPSG_WORLD_GEODETIC_SYSTEM);
+			ProjectionTransform wgs84Transform = projection
+					.getTransformation(printProjection);
 
 			BoundingBox boundingBox = tileMatrixSet.getBoundingBox();
 
@@ -426,7 +437,9 @@ public class ElevationTilesPngImportTest extends
 						if (algorithm == ElevationTilesAlgorithm.NEAREST_NEIGHBOR
 								|| (lat < maxLatitude && lon > minLongitude
 										&& lat > minLatitude && lon < maxLongitude)) {
-							TestCase.assertNotNull(elevation);
+							if (!allowNulls) {
+								TestCase.assertNotNull(elevation);
+							}
 						}
 					}
 					Double elevation = ElevationTilesPngTestUtils.getElevation(
@@ -436,7 +449,9 @@ public class ElevationTilesPngImportTest extends
 						System.out.print("   " + elevation);
 					}
 					if (algorithm == ElevationTilesAlgorithm.NEAREST_NEIGHBOR) {
-						TestCase.assertNotNull(elevation);
+						if (!allowNulls) {
+							TestCase.assertNotNull(elevation);
+						}
 					}
 				}
 				if (PRINT) {
@@ -450,7 +465,9 @@ public class ElevationTilesPngImportTest extends
 						System.out.print("   " + elevation);
 					}
 					if (algorithm == ElevationTilesAlgorithm.NEAREST_NEIGHBOR) {
-						TestCase.assertNotNull(elevation);
+						if (!allowNulls) {
+							TestCase.assertNotNull(elevation);
+						}
 					}
 				}
 				Double elevation = ElevationTilesPngTestUtils.getElevation(
@@ -460,7 +477,9 @@ public class ElevationTilesPngImportTest extends
 					System.out.print("   " + elevation);
 				}
 				if (algorithm == ElevationTilesAlgorithm.NEAREST_NEIGHBOR) {
-					TestCase.assertNotNull(elevation);
+					if (!allowNulls) {
+						TestCase.assertNotNull(elevation);
+					}
 				}
 
 				ElevationTileResults results = ElevationTilesPngTestUtils
@@ -481,7 +500,9 @@ public class ElevationTilesPngImportTest extends
 						if (PRINT) {
 							System.out.print("   " + elevations[y][x]);
 						}
-						TestCase.assertNotNull(elevation);
+						if (!allowNulls) {
+							TestCase.assertNotNull(elevation);
+						}
 					}
 				}
 
@@ -519,7 +540,9 @@ public class ElevationTilesPngImportTest extends
 								&& (row == 0 || column == 0)) {
 							TestCase.assertNull(elevation);
 						} else {
-							TestCase.assertNotNull(elevation);
+							if (!allowNulls) {
+								TestCase.assertNotNull(elevation);
+							}
 						}
 
 						elevation = ElevationTilesPngTestUtils.getElevation(
@@ -536,7 +559,9 @@ public class ElevationTilesPngImportTest extends
 										.getMatrixWidth() - 1)) {
 							TestCase.assertNull(elevation);
 						} else {
-							TestCase.assertNotNull(elevation);
+							if (!allowNulls) {
+								TestCase.assertNotNull(elevation);
+							}
 						}
 
 						elevation = ElevationTilesPngTestUtils.getElevation(
@@ -552,7 +577,9 @@ public class ElevationTilesPngImportTest extends
 								&& (row == tileMatrix.getMatrixHeight() - 1 || column == 0)) {
 							TestCase.assertNull(elevation);
 						} else {
-							TestCase.assertNotNull(elevation);
+							if (!allowNulls) {
+								TestCase.assertNotNull(elevation);
+							}
 						}
 
 						elevation = ElevationTilesPngTestUtils.getElevation(
@@ -569,7 +596,9 @@ public class ElevationTilesPngImportTest extends
 										.getMatrixWidth() - 1)) {
 							TestCase.assertNull(elevation);
 						} else {
-							TestCase.assertNotNull(elevation);
+							if (!allowNulls) {
+								TestCase.assertNotNull(elevation);
+							}
 						}
 
 						results = ElevationTilesPngTestUtils.getElevations(
@@ -599,7 +628,9 @@ public class ElevationTilesPngImportTest extends
 										System.out.print("   "
 												+ elevations[y][x]);
 									}
-									TestCase.assertNotNull(elevation);
+									if (!allowNulls) {
+										TestCase.assertNotNull(elevation);
+									}
 								}
 							}
 						}
@@ -628,8 +659,8 @@ public class ElevationTilesPngImportTest extends
 
 		for (ElevationTilesAlgorithm algorithm : ElevationTilesAlgorithm
 				.values()) {
-			Double elevation = ElevationTilesPngTestUtils.getElevation(geoPackage,
-					algorithm, latitude, longitude,
+			Double elevation = ElevationTilesPngTestUtils.getElevation(
+					geoPackage, algorithm, latitude, longitude,
 					ProjectionConstants.EPSG_WORLD_GEODETIC_SYSTEM);
 			if (PRINT) {
 				System.out.println(algorithm.name() + ": " + elevation);