+++ src/test/java/mil/nga/geopackage/test/features/columns/GeometryColumnsUtils.java
@@ -1,7 +1,6 @@
 package mil.nga.geopackage.test.features.columns;
 
 import java.sql.SQLException;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -527,7 +526,7 @@ public class GeometryColumnsUtils {
 			contents.setDataType(ContentsDataType.FEATURES);
 			contents.setIdentifier("test_contents");
 			contents.setDescription("");
-			contents.setLastChange(new Date());
+			// contents.setLastChange(new Date());
 			contents.setMinX(-180.0);
 			contents.setMinY(-90.0);
 			contents.setMaxX(180.0);